implementacion de la funcion de costo
esta funcion devuelve un costo asociado a cada operador. El de mayor
costo es el Bajar nivel (costo 50) y el de IrIngreso (costo 500) los
demas operadores tienen un costo de 10.
Se agrega el Panel de control para el ingreso de los datos.
Falta agregar la clase agente con los cambios para la eleccion de la
estategia de busqueda

+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/BajarNivel.java
@@ -163,7 +163,7 @@ public class BajarNivel extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(50);
     }
 
     /**
