+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/IrBaño.java
@@ -101,7 +101,7 @@ public class IrBaño extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(10);
     }
 
     /**
