+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/IrEscalera.java
@@ -125,7 +125,7 @@ public class IrEscalera extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(100);
     }
 
     /**
