+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/IrOficina.java
@@ -96,7 +96,7 @@ public class IrOficina extends SearchAction {
      */
     @Override
     public Double getCost() {
-        return new Double(0);
+        return new Double(10);
     }
 
     /**
