+++ TPIA2016/src/frsf/cidisi/exercise/tpia2016/search/actions/IrPasillo.java
@@ -112,7 +112,7 @@ public class IrPasillo extends SearchAction {
 	 */
 	@Override
 	public Double getCost() {
-		return new Double(0);
+		return new Double(10);
 	}
 
 	/**
